pyproject.toml
[tool.poetry]
name = "ACF Scoring API"
version = "0.1"
description = "Scoring service for ACF"
authors = ["zypl.ai"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "~0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras= ["asyncpg"], version = "2.0.28"}
asyncpg = "=0.29.0"
alembic = "=1.13.1"
python-multipart = "=0.0.9"
bcrypt= "~4.1.3"
aiohttp = "=3.9.5"
pyjwt = "~2.6.0"
XlsxWriter= "~3.1.8"

[tool.ruff]
select = ["ALL"]
ignore = ["E402", "E501", "EXE002", "D", "TCH"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
